origin/bonds/urls.py
from django.urls import include, path
from rest_framework import routers
from . import views, serializers
from rest_framework.authtoken.views import obtain_auth_token

router = routers.DefaultRouter()
router.register(r'bonds', views.BondView)

urlpatterns = [
    path('', include(router.urls)),
    path('api-token-auth/', obtain_auth_token, name='api_token_auth')
]
