origin/requirements.txt
certifi==2020.6.20
chardet==3.0.4
Django==2.2.13
djangorestframework==3.9.4
idna==2.10
leipy==0.1.0
pkg-resources==0.0.0
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.25.11
